pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-markdownify>=0.9.5",
    "django>=5.1.6",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "psycopg[binary,pool]>=3.2.5",
    "python-frontmatter>=1.1.0",
    "whitenoise>=6.9.0",
]
